move AccountsRequest to own file
