+++ src/lib/RpcApi.ts
@@ -3,7 +3,8 @@ import { BrowserDetection } from '@nimiq/utils';
 import { RootState } from '@/store';
 import { Store } from 'vuex';
 import Router from 'vue-router';
-import { AccountsRequest, RequestType } from './RequestTypes';
+import { RequestType } from './RequestTypes';
+import { AccountsRequest } from './AccountsRequest';
 import { RpcRequest, RpcResult } from './PublicRequestTypes';
 import { KeyguardCommand, KeyguardClient } from '@nimiq/keyguard-client';
 import { keyguardResponseRouter } from '@/router';
